// link2rail/projects/ekp/src/app/empty-wagon/api/generated/model/demand-wagon-type.ts
/**
 * Empty wagon API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface DemandWagonType { 
    /**
     * Wagon type code of the ordered wagon.
     */
    code?: string;
    /**
     * Wagon type number of the ordered wagon.
     */
    number: string;
    /**
     * The name of the Wagon type.
     */
    name?: string;
}